requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
Django==3.2.3
djangorestframework==3.12.4
flake8==3.9.2
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytz==2021.1
regex==2021.7.6
toml==0.10.2
